cleaned code on display.tsx by making it less monolithic
