+++ components/displays/forms/ViewDisplayForm.tsx
@@ -1,0 +1,40 @@
+import React, { useMemo } from "react";
+import { View } from "react-native";
+import { ViewFormMain } from "../../../utils/form";
+
+const ViewDisplayForm = (props: any) => {
+  // Chris todo: auxiliary data doesn't have relationship ids yet, so 'if(oldItem.focus_columns.cell_field==='relationship'){' does nothing yet
+
+  // create the appropriate schema for the form
+  let data: any = useMemo(() => {
+    let items: any = [];
+
+    if (props.schema && props.focus.data && props.auxiliary) {
+      props?.schema?.data?.forEach((oldItem: any) => {
+        let newItem = { ...oldItem, ...oldItem.focus_columns };
+        // if the attribute is 'relationship' we know that it is in the relationship table instead of being a column on the entity table.
+        if (oldItem.focus_columns.cell_field === "relationship") {
+          newItem.table = "relationships";
+          newItem.value = "(relationships)"; //'props.auxiliary.data to be filtered here (todo)'
+        } else {
+          newItem.table = "entities";
+          newItem.value = props.focus.data?.[0]?.[newItem.name_singular];
+        }
+
+        delete newItem.focus_columns;
+        delete newItem.auxiliary_columns;
+        items.push(newItem);
+      });
+
+      return items;
+    }
+  }, [props.schema, props.focus.data, props.auxiliary]);
+
+  return (
+    <View style={{ flexDirection: "column", maxHeight: 500 }}>
+      <ViewFormMain data={data} />
+    </View>
+  );
+};
+
+export default ViewDisplayForm;
