+++ components/displays/json/ViewJsonMain.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import { ScrollView, Text } from "react-native";
+
+const ViewJsonMain = (Input: any) => {
+  return (
+    // Placeholder. This will eventually allow for collapsing/expanding json nodes.
+    <ScrollView style={{ height: 400 }}>
+      <Text style={{ flex: 1 }}>{JSON.stringify(Input, null, 2)}</Text>
+    </ScrollView>
+  );
+};
+
+export default ViewJsonMain;
