+++ components/displays/maps/ViewDisplayMaps.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+import MapChart from "./maps";
+import { Text } from "react-native";
+
+const ViewDisplayMaps = (props: any) => {
+  return (
+    <>
+      <Text>This is maps page</Text>
+      <MapChart />
+    </>
+  );
+};
+
+export default ViewDisplayMaps;
