+++ utils/display.tsx
@@ -2,26 +2,21 @@
 // It is a 'perspective into the graph'.
 // You can view entities and their relationships in different displays (e.g. Table, list, calendar) and filters.
 
-import React, { ReactElement, memo, useMemo, useState } from "react";
-import { View, Text, Pressable } from "react-native";
-import { Link, useLocation, Route, Routes } from "react-router-dom";
-import { ViewListMain } from "./list";
-import { ViewTableMain, useTableColumns } from "./table";
-import { ViewJsonMain } from "./json";
-import { ViewIconMain } from "./icon";
-import { ViewPodsMain } from "./pods";
-import { ViewFormMain } from "./form";
+import React, { memo} from "react";
+import { View} from "react-native";
+import { Link, useLocation } from "react-router-dom";
 import {
-  ViewEntityAdd,
   useEntityArray,
   useEntitySingle,
   useEntitySchema,
 } from "./entity";
-import { ViewProcessesTabs } from "../components/entity/processTabs";
-import MyCalendar from "../components/displays/calendar";
-import Calendar from "react-calendar";
-import "react-calendar/dist/Calendar.css";
-import MapChart from "../components/displays/maps";
+import ViewDisplayCalendar from "../components/displays/calendar/ViewDisplayCalendar";
+import ViewDisplayForm from "../components/displays/forms/ViewDisplayForm";
+import ViewDisplayList from "../components/displays/list/ViewDisplayList";
+import ViewDisplayMaps from "../components/displays/maps/ViewDisplayMaps";
+import ViewDisplayPods from "../components/displays/pods/ViewDisplayPods";
+import ViewDisplayTable from "../components/displays/table/ViewDisplayTable";
+import ViewJsonMain from "../components/displays/json/ViewJsonMain";
 
 // Main
 
@@ -41,222 +36,7 @@ export const ViewDisplayMain = memo(({ component = "table" }: any) => {
 // Displays
 // These components transform the entity/schema/relationships data then provide them to the relevant component
 // (e.g. ViewDisplayTable transforms entities data, then provides them to the table component from table.js )
-// Chris is owning the below components.
 
-export const ViewDisplayList = (props: any) => {
-  const schema = props.schema;
-  const columns = useTableColumns(
-    schema.data?.map(
-      (x: { focus_columns: { name_singular: any } }) =>
-        x.focus_columns.name_singular
-    )
-  );
-  const auxiliary = props.auxiliary;
-  return (
-    <View style={{ flexDirection: "column" }}>
-      <ViewListMain columns={columns} data={auxiliary.data} />
-    </View>
-  );
-};
-
-export const ViewDisplayPods = (props: any) => {
-  const schema = props.schema;
-  const auxiliary = props.auxiliary;
-  const focus = props.focus;
-  const data = "make focus into an array and concat with aux";
-  return (
-    <View style={{ flexDirection: "column" }}>
-      <ViewPodsMain items={auxiliary} schema={schema.data}>
-        <ViewProcessesTabs />
-      </ViewPodsMain>
-    </View>
-  );
-};
-
-export const ViewDisplayForm = (props: any) => {
-  // Chris todo: auxiliary data doesn't have relationship ids yet, so 'if(oldItem.focus_columns.cell_field==='relationship'){' does nothing yet
-
-  // create the appropriate schema for the form
-  let data: any = useMemo(() => {
-    let items: any = [];
-
-    if (props.schema && props.focus.data && props.auxiliary) {
-      props?.schema?.data?.forEach((oldItem: any) => {
-        let newItem = { ...oldItem, ...oldItem.focus_columns };
-        // if the attribute is 'relationship' we know that it is in the relationship table instead of being a column on the entity table.
-        if (oldItem.focus_columns.cell_field === "relationship") {
-          newItem.table = "relationships";
-          newItem.value = "(relationships)"; //'props.auxiliary.data to be filtered here (todo)'
-        } else {
-          newItem.table = "entities";
-          newItem.value = props.focus.data?.[0]?.[newItem.name_singular];
-        }
-
-        delete newItem.focus_columns;
-        delete newItem.auxiliary_columns;
-        items.push(newItem);
-      });
-
-      return items;
-    }
-  }, [props.schema, props.focus.data, props.auxiliary]);
-
-  return (
-    <View style={{ flexDirection: "column", maxHeight: 500 }}>
-      <ViewFormMain data={data} />
-    </View>
-  );
-};
-
-export const ViewDisplayTable = (props: any) => {
-  const schema = props.schema;
-  const columns = useTableColumns(
-    schema.data?.map(
-      (x: { focus_columns: { name_singular: any } }) =>
-        x.focus_columns.name_singular
-    )
-  );
-  const auxiliary = props.auxiliary;
-  return (
-    <View style={{ flexDirection: "column" }}>
-      <ViewTableMain columns={columns} data={auxiliary.data} />
-    </View>
-  );
-};
-
-export const ViewDisplayCalendar = (props: any) => {
-  const schema = props.schema;
-  const auxiliary = props.auxiliary;
-
-  const [activeTab, setActiveTab] = useState(0);
-  const [value, onChange] = useState(new Date());
-
-  const handleTabPress = (index: number) => {
-    setActiveTab(index);
-  };
-
-  interface ICalendarEventBase {
-    start: Date;
-    end: Date;
-    title: string;
-    children?: ReactElement | null;
-  }
-
-  const eventNotes = useMemo(
-    () => (
-      <View style={{ marginTop: 3 }}>
-        <Text style={{ fontSize: 10, color: "white" }}>
-          {" "}
-          Phone number: 555-123-4567{" "}
-        </Text>
-        <Text style={{ fontSize: 10, color: "white" }}>
-          {" "}
-          Arrive 15 minutes early{" "}
-        </Text>
-      </View>
-    ),
-    []
-  );
-
-  const events: Array<ICalendarEventBase & { color?: string }> = [
-    {
-      title: "Team Meeting",
-      start: new Date(2023, 2, 15, 10, 0), // March 15, 2023, 10:00 AM
-      end: new Date(2023, 2, 15, 11, 0), // March 15, 2023, 11:00 AM
-      children: eventNotes,
-    },
-    {
-      title: "Project Review",
-      start: new Date(2023, 2, 17, 14, 30), // March 17, 2023, 2:30 PM
-      end: new Date(2023, 2, 17, 16, 0), // March 17, 2023, 4:00 PM
-      children: eventNotes,
-    },
-    {
-      title: "Conference Call",
-      start: new Date(2023, 2, 20, 11, 0), // March 20, 2023, 11:00 AM
-      end: new Date(2023, 2, 20, 12, 0), // March 20, 2023, 12:00 PM
-      children: eventNotes,
-    },
-    {
-      title: "Team Building Workshop",
-      start: new Date(2023, 2, 22, 9, 30), // March 22, 2023, 9:30 AM
-      end: new Date(2023, 2, 22, 12, 30), // March 22, 2023, 12:30 PM
-      children: eventNotes,
-    },
-    {
-      title: "Product Demo",
-      start: new Date(2023, 2, 25, 15, 0), // March 25, 2023, 3:00 PM
-      end: new Date(2023, 2, 25, 16, 0), // March 25, 2023, 4:00 PM
-      children: eventNotes,
-    },
-    // Add more random events as needed
-  ];
-
-  const tabs = [
-    { tab: "View Calendar", component: <MyCalendar myEventsList={events} /> },
-    {
-      tab: "Add events",
-      component: <Calendar onChange={onChange} value={value} />,
-    },
-  ];
-
-  return (
-    <>
-      <View
-        style={{
-          flexDirection: "row",
-          marginBottom: 10,
-          marginHorizontal: 12,
-        }}
-      >
-        {tabs.map((content, index) => (
-          <Pressable
-            key={index}
-            style={{
-              padding: 10,
-              borderTopWidth: 1,
-              borderRightWidth: 1,
-              borderLeftWidth: 1,
-              borderTopRightRadius: 5,
-              borderTopLeftRadius: 5,
-              borderColor: activeTab === index ? "black" : "transparent",
-              backgroundColor:
-                activeTab === index ? "lightblue" : "transparent",
-            }}
-            onPress={() => handleTabPress(index)}
-          >
-            <Text style={{ fontWeight: "bold" }}>{content.tab}</Text>
-          </Pressable>
-        ))}
-      </View>
-
-      <View>{tabs[activeTab].component}</View>
-    </>
-  );
-};
-
-export const ViewDisplayMaps = (props: any) => {
-  return (
-    <>
-      <Text>This is maps page</Text>
-      <MapChart />
-    </>
-  );
-}
-
-export const ViewDisplayJson = (props: any) => {
-  const schema = props.schema;
-  const auxiliary = props.auxiliary;
-  const columns = useTableColumns(
-    schema.data?.map((x: any) => x.focus_columns.name_singular)
-  );
-
-  return (
-    <View style={{ flexDirection: "column" }}>
-      <ViewJsonMain schema={schema} auxiliary={auxiliary} columns={columns} />
-    </View>
-  );
-};
 
 // Components
 
@@ -311,29 +91,6 @@ export const optionsDisplayActions = [
   },
   // {title:'Display',iconName:'eye',iconSource:'Feather', description: "change the display mode between Calendar, Table, List etc."},// in case we want to make the mode/display choices (calendar, table etc.) within these tabs instead of on their own
 ];
-
-export const ViewDisplayActions = ({}: any) => {
-  return (
-    <View style={{ flexDirection: "row" }}>
-      <View style={{ flexDirection: "row" }}>
-        {optionsDisplayActions?.map((x, i) => (
-          <Link key={i} style={{ padding: 5 }} to={x.title.toLowerCase()}>
-            {/* {x.title} */}
-            <ViewIconMain
-              name={x.iconName}
-              source={x.iconSource}
-              color={"black"}
-              size={24}
-            />
-          </Link>
-        ))}
-      </View>
-      <Routes>
-        <Route path="/add" element={<ViewEntityAdd />} />
-      </Routes>
-    </View>
-  );
-};
 
 // Tabs
